Se agregaron las validaciones de registro y modificacion de genero y productos, ademas de las validaciones de campos en el carrito de compra

+++ database/migrations/2019_08_03_205313_create_pedidos_table.php
@@ -18,8 +18,8 @@ class CreatePedidosTable extends Migration
             $table->bigInteger('referencia');
             $table->date('fecha');
             $table->string('nombre_cliente');
-            $table->integer('telefono');
-            $table->integer('cedula');
+            $table->bigInteger('telefono');
+            $table->bigInteger('cedula');
             $table->string('departamento');
             $table->string('ciudad');
             $table->string('barrio');
